+++ connect-for-commerce-scope-server/src/main/java/to/be/renamed/EcomConnectScope.java
@@ -58,6 +58,10 @@ public class EcomConnectScope {
         }
     }
 
+    public static EcomConnectScope create(SpecialistsBroker broker) {
+        return new EcomConnectScope(broker);
+    }
+
     public SpecialistsBroker getBroker() {
         return broker;
     }